Added a 'service layer', 'Exception layer', 'Cofig layer' and a bunch of 'Request' DTOs

+++ src/main/java/com/codewithdaeshaun/store/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.codewithdaeshaun.store.repositories;
 import com.codewithdaeshaun.store.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Byte> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 }
